Split of ExperimentRunnter into a Preparer and a Runner
